feat(telemetry): classify 401/403/404 funnel errors and share reason union
